tracker.csv and schedule.csv had strayed from their descriptions in outline.txt. I reworked them both to be more in line with their descriptions. tracker.csv now only tracks when a bundle of jobs has been completed (now called a bundle instead of an interval) and schedule.csv only tracks which jobs correspond to which schedule--it is read-only. scheduler.py has been updated to account for this.
scheduler.py now updates a whole bundle at a time instead of one job at a time. add_task() became add_bundle(). add_bundle() still checks for if a job exists on todo.txt--this is in case a bundle is not completed during a week.

I just realized as I was writing the last line that scheduler.py will recreate an entire bundle each monday if it's not completed by friday. I will need to work on that, adding a way to check if a task is currently in work. This could be accomplished by a tracker file or by checking which IDs and thus which bundles are present in todo.txt.

scheduler.py now will read all types of tracker file, so the shop infra and the cars both have their own tracker file now.